[refactor] #21 service layer 분리
